+++ docs/CHANGELOG.md
@@ -45,6 +45,7 @@
 - `[Radio]` Fixed alignment in datagrid radio button. ([#8349](https://github.com/infor-design/enterprise/issues/8349))
 - `[Searchfield]` Fixed cut-off searchfield button. ([#8333](https://github.com/infor-design/enterprise/issues/8333))
 - `[Searchfield]` Fixed clear button misalignment in RTL. ([#8342](https://github.com/infor-design/enterprise/issues/8342))
+- `[Tabs]` Fixed error in tab popup. ([#8596](https://github.com/infor-design/enterprise/issues/8596))
 - `[Tabs]` Added guards on possible undefined objects. ([#8419](https://github.com/infor-design/enterprise/issues/8419))
 - `[Tabs]` Fixed `beforeactivated` event not cancelling activation of tabs properly. ([NG#1578](https://github.com/infor-design/enterprise-ng/issues/1578))
 - `[TabsHeader]` Fixed tooltip not showing properly when there is an ellipsis in tabs. ([#8446](https://github.com/infor-design/enterprise/issues/8446))